[feat] Admin 공방 평균 별점 구하는 기능 구현

+++ FinalProject_GotGongBang/src/main/java/com/spring/gotgongbang/admin/model/InterAdminDAO.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Map;
 
 import com.spring.gotgongbang.craft.model.CraftVO;
+import com.spring.gotgongbang.member.model.ReviewVO;
 
 
 public interface InterAdminDAO {
@@ -46,7 +47,7 @@ public interface InterAdminDAO {
 	List<CraftVO> getCraftListWithPaging(Map<String, String> paraMap);
 
 	//공방 평균 별점 가져오기
-	int craftRating(String craft_num_pk);
+	long craftRating(String craft_num_pk);
